src/main/java/com/zipcodewilmington/singlylinkedlist/Node.java
package com.zipcodewilmington.singlylinkedlist;

public class Node<T> {

    private T obj;
    private Node next;
    private int index = 0;

    public Node(T obj, Node next){
        this.obj = obj;
        this.next = next;
    }

    public T getObj() {
        return obj;
    }

    public void setObj(T obj) {
        this.obj = obj;
    }

    public Node getNext() {
        return next;
    }

    public void setNext(Node next) {
        this.next = next;
    }

    public int getIndex() {
        return index;
    }

    public void setIndex(int index) {
        this.index = index;
    }
}
